updating track playing time logic was moved from PlayerActivity to PlayerViewModel
