+++ app/src/main/java/bes/max/playlistmaker/ui/PlayerViewModel.kt
@@ -1,5 +1,7 @@
 package bes.max.playlistmaker.ui
 
+import android.os.Handler
+import android.os.Looper
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import bes.max.playlistmaker.domain.Player
@@ -11,7 +13,10 @@ class PlayerViewModel(val track: Track) : ViewModel() {
 
     private val player: Player = Player()
     val playerState = player.playerState
-    val playingTime = MutableLiveData<Int>(0) //TODO change this value and update ui with seconds
+    val playingTime =
+        MutableLiveData<String>("00:00") //TODO change this value and update ui with seconds
+    private val handler = Handler(Looper.getMainLooper())
+    private val timerRunnable = Runnable { updateTimer() }
 
     init {
         player.createNewMediaPlayer()
@@ -31,10 +36,12 @@ class PlayerViewModel(val track: Track) : ViewModel() {
 
             else -> {}
         }
+        updateTimer()
     }
 
     fun releasePlayer() {
         player.releasePlayer()
+        updateTimer()
     }
 
     fun getCurrentPlayerPositionAsNumber(): Int {
@@ -46,5 +53,31 @@ class PlayerViewModel(val track: Track) : ViewModel() {
         return SimpleDateFormat("mm:ss", Locale.getDefault()).format(time)
     }
 
+    private fun updateTimer() {
+        val formattedText = formatIntToFormattedTimeText(
+            getCurrentPlayerPositionAsNumber()
+        )
+        when (playerState.value) {
+            Player.PlayerState.STATE_PLAYING -> {
+                playingTime.value =
+                    formattedText
+                handler.postDelayed(timerRunnable, TIMER_UPDATE_RATE)
+            }
+
+            Player.PlayerState.STATE_PAUSED -> {
+                handler.removeCallbacks(timerRunnable)
+            }
+
+            else -> {
+                handler.removeCallbacks(timerRunnable)
+                playingTime.value = DEFAULT_TIMER_TIME
+            }
+        }
+    }
+
+    companion object {
+        private const val TIMER_UPDATE_RATE = 500L
+        private const val DEFAULT_TIMER_TIME = "00:00"
+    }
 
 }
